WIP, moving files v2 to core
