+++ app/api/core/application/specs/EntitiesService.spec.ts
@@ -10,7 +10,7 @@ import { SettingsDataSourceFactory } from 'api/core/infrastructure/factories/Set
 import { TemplatesDataSourceFactory } from 'api/core/infrastructure/factories/TemplatesDataSourceFactory';
 import { getConnection } from 'api/core/infrastructure/mongodb/common/getConnectionForCurrentTenant';
 import { MongoMultiLanguageEntityDataSource } from 'api/entities.v2/database/MongoMultiLanguageEntityDataSource';
-import { FileSystemStorage } from 'api/files.v2/infrastructure/FileSystemStorage';
+import { FileSystemStorage } from 'api/core/infrastructure/files/FileSystemStorage';
 import { TestUtils } from 'api/common.v2/utils/Test';
 import { tenants } from 'api/tenants';
 import { EventsBus } from 'api/core/libs/eventsbus';
