+++ app/api/core/infrastructure/factories/FilesDataSourceFactory.ts
@@ -1,7 +1,7 @@
 import { getConnection } from 'api/core/infrastructure/mongodb/common/getConnectionForCurrentTenant';
 import { MongoTransactionManager } from 'api/core/infrastructure/mongodb/common/MongoTransactionManager';
-import { MongoFilesDataSource } from '../../../files.v2/database/MongoFilesDataSource';
-import { FileStorageFactory } from '../../../files.v2/infrastructure/FileStorageFactory';
+import { MongoFilesDataSource } from '../mongodb/files/MongoFilesDataSource';
+import { FileStorageFactory } from '../files/FileStorageFactory';
 
 export class FilesDataSourceFactory {
   static default(transactionManager: MongoTransactionManager) {
