+++ app/api/core/infrastructure/mongodb/files/specs/MongoFilesDataSource.spec.ts
@@ -1,14 +1,14 @@
 import { TransactionManagerFactory } from 'api/core/infrastructure/factories/TransactionManagerFactory';
+import { FileStorageFactory } from 'api/core/infrastructure/files/FileStorageFactory';
 import { getConnection } from 'api/core/infrastructure/mongodb/common/getConnectionForCurrentTenant';
-import { FileStorageFactory } from 'api/files.v2/infrastructure/FileStorageFactory';
 import { DiskFile } from 'api/files.v2/model/DiskFile';
 import { Document } from 'api/files.v2/model/Document';
 import { FileNotFound } from 'api/files.v2/model/errors';
 import { ProcessedDocument } from 'api/files.v2/model/ProcessedDocument';
+import { FileBuilder } from 'api/files.v2/specs/FileBuilder';
 import { elasticTesting } from 'api/utils/elastic_testing';
 import { getFixturesFactory } from 'api/utils/fixturesFactory';
 import { testingEnvironment } from 'api/utils/testingEnvironment';
-import { FileBuilder } from 'api/files.v2/specs/FileBuilder';
 import { MongoFilesDataSource } from '../MongoFilesDataSource';
 
 const f = getFixturesFactory();
