+++ app/api/core/infrastructure/mongodb/files/specs/MongoSegmentationBuilder.ts
