+++ app/api/csv.v2/jobs/specs/CsvExtractUploadedZipJob.spec.ts
@@ -7,8 +7,8 @@ import { tenants } from 'api/tenants/tenantContext';
 import { testingEnvironment } from 'api/utils/testingEnvironment';
 import { TransactionManagerFactory } from 'api/core/infrastructure/factories/TransactionManagerFactory';
 import { DefaultCsvImportsDataSource } from 'api/csv.v2/database/data_source_defaults';
-import { FileSystemStorage } from 'api/files.v2/infrastructure/FileSystemStorage';
-import { PathManager } from 'api/files.v2/infrastructure/PathManager';
+import { FileSystemStorage } from 'api/core/infrastructure/files/FileSystemStorage';
+import { PathManager } from 'api/core/infrastructure/files/PathManager';
 import { DiskFile } from 'api/files.v2/model/DiskFile';
 import { CsvImportDomain, CsvImportStatus } from 'api/csv.v2/model/CsvImport';
 import { createTestingZip } from 'api/csv/specs/helpers';
