+++ app/api/paragraphExtraction/application/PXGetExtractorStatuses.ts
@@ -2,7 +2,7 @@ import { LanguageISO6391 } from 'shared/types/commonTypes';
 
 import { UseCase } from 'api/core/libs/UseCase';
 import { SettingsDataSource } from 'api/core/application/contracts/SettingsDataSource';
-import { FilesDataSource } from 'api/files.v2/contracts/FilesDataSource';
+import { FilesDataSource } from 'api/core/application/contracts/FilesDataSource';
 
 import {
   GetExtractorStatusesInput,
